Remove worker threads from alerter, inline matched processing
Worker threads added complexity without meaningful benefit for I/O-bound
alerter work. This inlines all matched processing directly into app.js
with a single PromiseQueue, giving better geocoder cache hit rates,
simpler backpressure, and no MessageChannel overhead.

- Remove worker_threads infrastructure from app.js
- Delete matchedWorker.js, inline logic into app.js
- Remove setImmediate wrappers from controllers so PromiseQueue properly
  gates concurrency (monster, raid, gym, pokestop, pokestop_lure)
- Add circuit breaker to cachingGeocoder and tileserverPregen
- Fix lint: remove unused S2 imports, == to ===, eslint-disable for
  no-continue and no-nested-ternary
- Add .eslintrc config
