+++ alerter/.eslintrc
@@ -1,0 +1,23 @@
+{
+    "parserOptions": { "ecmaVersion": 12 },
+    "extends": "airbnb-base",
+    "rules": {
+        "indent": [ "error", "tab", { "SwitchCase": 1 } ],
+        "semi": [ "error", "never" ],
+        "import/no-dynamic-require": [ "off" ],
+        "global-require": [ "off" ],
+        "no-tabs": [ "off" ],
+        "array-callback-return": [ "off" ],
+        "max-len": [ "off" ],
+        "eol-last": [ "off" ],
+        "consistent-return": [ "off" ],
+        "no-restricted-syntax": [ "off" ],
+        "no-mixed-operators": [ "off" ],
+        "no-return-assign": ["off"],
+        "no-await-in-loop": [ "off" ],
+        "no-plusplus": [ "off" ],
+        "no-throw-literal": ["off"],
+        "no-param-reassign": ["off"],
+        "no-bitwise": ["error", { "allow": ["~"] }]
+    }
+}
