+++ alerter/src/controllers/monster.js
@@ -1,8 +1,6 @@
 const geoTz = require('geo-tz')
 const moment = require('moment-timezone')
 require('moment-precise-range-plugin')
-const { S2 } = require('s2-geometry')
-const S2ts = require('nodes2ts')
 const Controller = require('./controller')
 
 class Monster extends Controller {
@@ -143,7 +141,7 @@ class Monster extends Controller {
 							caps.push(stats.cap)
 						}
 						for (const cap of caps) {
-							if (!capsConsidered.includes(cap)) continue
+							if (!capsConsidered.includes(cap)) continue // eslint-disable-line no-continue
 							if (stats.rank && stats.rank < best[cap].rank) {
 								best[cap].rank = stats.rank
 								best[cap].cp = stats.cp || 0
@@ -154,8 +152,13 @@ class Monster extends Controller {
 
 						if (!stats.evolution && this.config.pvp.pvpEvolutionDirectTracking && stats.rank && stats.cp && stats.pokemon !== data.pokemon_id && stats.rank <= this.config.pvp.pvpFilterMaxRank && stats.cp >= minCp) {
 							const newEvo = {
-								rank: stats.rank, percentage: stats.percentage, pokemon: stats.pokemon, form: stats.form || 0, level: stats.level, cp: stats.cp,
-								caps: stats.capped ? capsConsidered.filter((x) => x >= stats.cap) : (stats.cap && capsConsidered.includes(stats.cap) ? [stats.cap] : null),
+								rank: stats.rank,
+								percentage: stats.percentage,
+								pokemon: stats.pokemon,
+								form: stats.form || 0,
+								level: stats.level,
+								cp: stats.cp,
+								caps: stats.capped ? capsConsidered.filter((x) => x >= stats.cap) : (stats.cap && capsConsidered.includes(stats.cap) ? [stats.cap] : null), // eslint-disable-line no-nested-ternary
 							}
 							if (data.pvpEvolutionData[stats.pokemon] && data.pvpEvolutionData[stats.pokemon][league]) {
 								data.pvpEvolutionData[stats.pokemon][league].push(newEvo)
@@ -222,368 +225,366 @@ class Monster extends Controller {
 			}
 
 			// From here, same as handle() post-whoCares: data enrichment + message creation
-			setImmediate(async () => {
-				try {
-					let discordCacheBad = true
+			try {
+				let discordCacheBad = true
+				whoCares.forEach((cares) => {
+					if (!this.isRateLimited(cares.id)) discordCacheBad = false
+				})
+
+				if (discordCacheBad) {
 					whoCares.forEach((cares) => {
-						if (!this.isRateLimited(cares.id)) discordCacheBad = false
+						this.log.verbose(`${logReference}: [matched] Not creating monster alert (Rate limit) for ${cares.type} ${cares.id} ${cares.name}`)
 					})
+					return []
+				}
 
-					if (discordCacheBad) {
-						whoCares.forEach((cares) => {
-							this.log.verbose(`${logReference}: [matched] Not creating monster alert (Rate limit) for ${cares.type} ${cares.id} ${cares.name}`)
+				const consolidatedAlerts = []
+				for (const alert of whoCares) {
+					let existingAlert = consolidatedAlerts.find((x) => x.id === alert.id)
+					if (!existingAlert) {
+						existingAlert = { ...alert, filters: [] }
+						consolidatedAlerts.push(existingAlert)
+					}
+					if (alert.pvp_ranking_worst < 4096) {
+						existingAlert.filters.push({
+							pvp_ranking_league: alert.pvp_ranking_league,
+							pvp_ranking_worst: alert.pvp_ranking_worst,
+							pvp_ranking_cap: alert.pvp_ranking_cap,
 						})
-						return []
 					}
+				}
 
-					const consolidatedAlerts = []
-					for (const alert of whoCares) {
-						let existingAlert = consolidatedAlerts.find((x) => x.id === alert.id)
-						if (!existingAlert) {
-							existingAlert = { ...alert, filters: [] }
-							consolidatedAlerts.push(existingAlert)
-						}
-						if (alert.pvp_ranking_worst < 4096) {
-							existingAlert.filters.push({
-								pvp_ranking_league: alert.pvp_ranking_league,
-								pvp_ranking_worst: alert.pvp_ranking_worst,
-								pvp_ranking_cap: alert.pvp_ranking_cap,
-							})
-						}
+				if (data.display_pokemon_id && data.display_pokemon_id !== data.pokemon_id) {
+					data.display_form ??= 0
+					const displayMonster = this.GameData.monsters[`${data.display_pokemon_id}_${data.display_form}`] || this.GameData.monsters[`${data.display_pokemon_id}_0`]
+					if (displayMonster) {
+						data.disguisePokemonNameEng = displayMonster.name
+						if (displayMonster.form) data.disguideFormNameEng = displayMonster.form.name
 					}
+				}
 
-					if (data.display_pokemon_id && data.display_pokemon_id !== data.pokemon_id) {
-						data.display_form ??= 0
-						const displayMonster = this.GameData.monsters[`${data.display_pokemon_id}_${data.display_form}`] || this.GameData.monsters[`${data.display_pokemon_id}_0`]
-						if (displayMonster) {
-							data.disguisePokemonNameEng = displayMonster.name
-							if (displayMonster.form) data.disguideFormNameEng = displayMonster.form.name
-						}
+				if (this.imgUicons) data.imgUrl = await this.imgUicons.pokemonIcon(data.pokemon_id, data.form, 0, data.gender, data.costume, 0, data.shinyPossible && this.config.general.requestShinyImages) || this.config.fallbacks?.imgUrl
+				if (this.imgUiconsAlt) data.imgUrlAlt = await this.imgUiconsAlt.pokemonIcon(data.pokemon_id, data.form, 0, data.gender, data.costume, 0, data.shinyPossible && this.config.general.requestShinyImages) || this.config.fallbacks?.imgUrl
+				if (this.stickerUicons) data.stickerUrl = await this.stickerUicons.pokemonIcon(data.pokemon_id, data.form, 0, data.gender, data.costume, 0, data.shinyPossible && this.config.general.requestShinyImages)
+
+				const geoResult = await this.getAddress({ lat: data.latitude, lon: data.longitude })
+
+				const jobs = []
+
+				require('./common/nightTime').setNightTime(data, disappearTime, this.config)
+
+				data.intersection = await this.obtainIntersection(data)
+
+				if (data.seen_type) {
+					switch (data.seen_type) {
+						case 'nearby_stop': data.seenType = 'pokestop'; break
+						case 'nearby_cell': data.seenType = 'cell'; break
+						case 'lure':
+						case 'lure_wild': data.seenType = 'lure'; break
+						case 'lure_encounter':
+						case 'encounter':
+						case 'wild': data.seenType = data.seen_type; break
+						default: break
 					}
+				} else if (data.pokestop_id === 'None' && data.spawnpoint_id === 'None') data.seenType = 'cell'
+				else if (data.pokestop_id === 'None') data.seenType = encountered ? 'encounter' : 'wild'
+				else data.seenType = 'pokestop'
 
-					if (this.imgUicons) data.imgUrl = await this.imgUicons.pokemonIcon(data.pokemon_id, data.form, 0, data.gender, data.costume, 0, data.shinyPossible && this.config.general.requestShinyImages) || this.config.fallbacks?.imgUrl
-					if (this.imgUiconsAlt) data.imgUrlAlt = await this.imgUiconsAlt.pokemonIcon(data.pokemon_id, data.form, 0, data.gender, data.costume, 0, data.shinyPossible && this.config.general.requestShinyImages) || this.config.fallbacks?.imgUrl
-					if (this.stickerUicons) data.stickerUrl = await this.stickerUicons.pokemonIcon(data.pokemon_id, data.form, 0, data.gender, data.costume, 0, data.shinyPossible && this.config.general.requestShinyImages)
-
-					const geoResult = await this.getAddress({ lat: data.latitude, lon: data.longitude })
-
-					const jobs = []
-
-					require('./common/nightTime').setNightTime(data, disappearTime, this.config)
-
-					data.intersection = await this.obtainIntersection(data)
-
-					if (data.seen_type) {
-						switch (data.seen_type) {
-							case 'nearby_stop': data.seenType = 'pokestop'; break
-							case 'nearby_cell': data.seenType = 'cell'; break
-							case 'lure':
-							case 'lure_wild': data.seenType = 'lure'; break
-							case 'lure_encounter':
-							case 'encounter':
-							case 'wild': data.seenType = data.seen_type; break
-							default: break
-						}
-					} else if (data.pokestop_id === 'None' && data.spawnpoint_id === 'None') data.seenType = 'cell'
-					else if (data.pokestop_id === 'None') data.seenType = encountered ? 'encounter' : 'wild'
-					else data.seenType = 'pokestop'
-
-					if (data.seenType === 'cell' && !data.cell_coords) {
-						const { S2 } = require('s2-geometry')
-						const S2ts = require('nodes2ts')
-						const areaCellKey = S2.latLngToKey(data.latitude, data.longitude, 15)
-						const areaCellId = S2.keyToId(areaCellKey)
-						const s2cell = new S2ts.S2Cell(new S2ts.S2CellId(areaCellId))
-						data.cell_coords = []
-						for (let i = 0; i <= 3; i++) {
-							const vertex = S2ts.S2LatLng.fromPoint(s2cell.getVertex(i))
-							data.cell_coords.push([parseFloat(vertex.latDegrees), parseFloat(vertex.lngDegrees)])
-						}
+				if (data.seenType === 'cell' && !data.cell_coords) {
+					const { S2 } = require('s2-geometry')
+					const S2ts = require('nodes2ts')
+					const areaCellKey = S2.latLngToKey(data.latitude, data.longitude, 15)
+					const areaCellId = S2.keyToId(areaCellKey)
+					const s2cell = new S2ts.S2Cell(new S2ts.S2CellId(areaCellId))
+					data.cell_coords = []
+					for (let i = 0; i <= 3; i++) {
+						const vertex = S2ts.S2LatLng.fromPoint(s2cell.getVertex(i))
+						data.cell_coords.push([parseFloat(vertex.latDegrees), parseFloat(vertex.lngDegrees)])
 					}
+				}
 
-					await this.getStaticMapUrl(
-						logReference,
-						data,
-						'monster',
-						['pokemon_id', 'latitude', 'longitude', 'form', 'costume', 'imgUrl', 'imgUrlAlt', 'style'],
-						['pokemon_id', 'display_pokemon_id', 'latitude', 'longitude', 'verified', 'costume', 'form', 'pokemonId', 'generation', 'weather', 'confirmedTime', 'shinyPossible', 'seenType', 'seen_type', 'cell_coords', 'imgUrl', 'imgUrlAlt', 'nightTime', 'duskTime', 'dawnTime', 'style'],
-					)
-					data.staticmap = data.staticMap // deprecated
+				await this.getStaticMapUrl(
+					logReference,
+					data,
+					'monster',
+					['pokemon_id', 'latitude', 'longitude', 'form', 'costume', 'imgUrl', 'imgUrlAlt', 'style'],
+					['pokemon_id', 'display_pokemon_id', 'latitude', 'longitude', 'verified', 'costume', 'form', 'pokemonId', 'generation', 'weather', 'confirmedTime', 'shinyPossible', 'seenType', 'seen_type', 'cell_coords', 'imgUrl', 'imgUrlAlt', 'nightTime', 'duskTime', 'dawnTime', 'style'],
+				)
+				data.staticmap = data.staticMap // deprecated
 
-					// Weather forecast
-					const { nextHourTimestamp } = this.weatherData.getWeatherTimes()
-					if (this.config.weather.enableWeatherForecast && data.disappear_time > nextHourTimestamp) {
-						const weatherCellId = this.weatherData.getWeatherCellId(data.latitude, data.longitude)
-						const weatherForecast = await this.weatherData.getWeatherForecast(weatherCellId)
+				// Weather forecast
+				const { nextHourTimestamp } = this.weatherData.getWeatherTimes()
+				if (this.config.weather.enableWeatherForecast && data.disappear_time > nextHourTimestamp) {
+					const weatherCellId = this.weatherData.getWeatherCellId(data.latitude, data.longitude)
+					const weatherForecast = await this.weatherData.getWeatherForecast(weatherCellId)
 
-						let pokemonShouldBeBoosted = false
-						let pokemonWillBeBoosted = false
-						const currentBoostedTypes = weatherForecast.current ? this.GameData.utilData.weatherTypeBoost[weatherForecast.current] : []
-						const forecastBoostedTypes = weatherForecast.next ? this.GameData.utilData.weatherTypeBoost[weatherForecast.next] : []
-						if (weatherForecast.current > 0 && currentBoostedTypes.filter((boostedType) => data.types.includes(boostedType)).length > 0) pokemonShouldBeBoosted = true
-						if (weatherForecast.next > 0 && ((data.weather > 0 && weatherForecast.next !== data.weather) || (weatherForecast.current > 0 && weatherForecast.next !== weatherForecast.current) || (pokemonShouldBeBoosted && data.weather === 0))) {
-							const weatherChangeTime = moment((data.disappear_time - (data.disappear_time % 3600)) * 1000)
-								.tz(geoTz.find(data.latitude, data.longitude)[0].toString())
-								.format(this.config.locale.time)
-								.slice(0, -3)
-							pokemonWillBeBoosted = forecastBoostedTypes.filter((boostedType) => data.types.includes(boostedType)).length > 0 ? 1 : 0
-							if (data.weather > 0 && !pokemonWillBeBoosted || data.weather === 0 && pokemonWillBeBoosted) {
-								weatherForecast.current = data.weather > 0 ? data.weather : weatherForecast.current
-								if (pokemonShouldBeBoosted && data.weather === 0) {
-									data.weatherCurrent = 0
-								} else {
-									data.weatherCurrent = weatherForecast.current
-								}
-								data.weatherChangeTime = weatherChangeTime
-								data.weatherNext = weatherForecast.next
-							}
-						}
-					}
-
-					// Future event processing
-					const event = this.eventParser.eventChangesSpawn(moment().unix(), data.disappear_time, data.latitude, data.longitude)
-					if (event) {
-						data.futureEvent = true
-						data.futureEventTime = event.time
-						data.futureEventName = event.name
-						data.futureEventTrigger = event.reason
-					}
-
-					// Lookup pokestop name if needed
-					if (this.config.general.populatePokestopName && !data.pokestopName && data.pokestop_id && this.scannerQuery) {
-						data.pokestopName = this.escapeJsonString(await this.scannerQuery.getPokestopName(data.pokestop_id))
-					}
-
-					for (const cares of consolidatedAlerts) {
-						this.log.debug(`${logReference}: [matched] Creating monster alert for ${cares.id} ${cares.name} ${cares.type} ${cares.language} ${cares.template}`)
-
-						const rateLimitTtr = this.getRateLimitTimeToRelease(cares.id)
-						if (rateLimitTtr) {
-							this.log.verbose(`${logReference}: [matched] Not creating monster alert (Rate limit) for ${cares.type} ${cares.id} ${cares.name} Time to release: ${rateLimitTtr}`)
-							continue
-						}
-
-						const language = cares.language || this.config.general.locale
-						const translator = this.translatorFactory.Translator(language)
-						let [platform] = cares.type.split(':')
-						if (platform === 'webhook') platform = 'discord'
-
-						data.name = translator.translate(monster.name)
-						data.formName = translator.translate(monster.form.name)
-						if (data.disguisePokemonNameEng) data.disguisePokemonName = translator.translate(data.disguisePokemonNameEng)
-						if (data.disguiseFormNameEng) data.disguiseFormName = translator.translate(data.disguiseFormNameEng)
-
-						data.formNormalisedEng = data.formNameEng === 'Normal' ? '' : data.formNameEng
-						data.formNormalised = translator.translate(data.formNormalisedEng)
-
-						data.fullNameEng = data.nameEng.concat(data.formNormalisedEng ? ' ' : '', data.formNormalisedEng)
-						data.fullName = data.name.concat(data.formNormalised ? ' ' : '', data.formNormalised)
-
-						data.genderData = {
-							name: translator.translate(data.genderDataEng.name),
-							emoji: translator.translate(this.emojiLookup.lookup(data.genderDataEng.emoji, platform)),
-						}
-						data.genderName = data.genderData.name
-						data.genderEmoji = data.genderData.emoji
-						data.shinyPossibleEmoji = data.shinyPossible ? translator.translate(this.emojiLookup.lookup('shiny', platform)) : ''
-						data.rarityName = translator.translate(data.rarityNameEng)
-						data.sizeName = translator.translate(data.sizeNameEng)
-						data.quickMoveName = encountered && this.GameData.moves[data.quickMoveId] ? translator.translate(this.GameData.moves[data.quickMoveId].name) : ''
-						data.quickMoveEmoji = this.GameData.moves[data.quickMoveId] && this.GameData.utilData.types[this.GameData.moves[data.quickMoveId].type] ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.types[this.GameData.moves[data.quickMoveId].type].emoji, platform)) : ''
-						data.chargeMoveName = encountered && this.GameData.moves[data.chargeMoveId] ? translator.translate(this.GameData.moves[data.chargeMoveId].name) : ''
-						data.chargeMoveEmoji = this.GameData.moves[data.chargeMoveId] && this.GameData.utilData.types[this.GameData.moves[data.chargeMoveId].type] ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.types[this.GameData.moves[data.chargeMoveId].type].emoji, platform)) : ''
-						data.boosted = !!data.weather
-						data.generationName = translator.translate(data.generationNameEng)
-						data.boostWeatherId = data.weather ? data.weather : ''
-						data.boostWeatherName = data.weather ? translator.translate(this.GameData.utilData.weather[data.weather].name) : ''
-						data.boostWeatherEmoji = data.weather ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weather].emoji, platform)) : ''
-						data.gameWeatherId = this.GameData.utilData.weather[currentCellWeather] ? currentCellWeather : ''
-						data.gameWeatherName = this.GameData.utilData.weather[currentCellWeather] ? translator.translate(this.GameData.utilData.weather[currentCellWeather].name) : ''
-						data.gameWeatherEmoji = this.GameData.utilData.weather[currentCellWeather] ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[currentCellWeather].emoji, platform)) : ''
-						data.formname = data.formName // deprecated
-						data.quickMove = data.quickMoveName // deprecated
-						data.chargeMove = data.chargeMoveName // deprecated
-						data.move1emoji = data.quickMoveEmoji // deprecated
-						data.move2emoji = data.chargeMoveEmoji // deprecated
-						data.boost = data.boostWeatherName // deprecated
-						data.boostemoji = data.boostWeatherEmoji // deprecated
-						data.gameweather = data.gameWeatherName // deprecated
-						data.gameweatheremoji = data.gameWeatherEmoji // deprecated
-
-						if (data.weatherNext) {
-							if (!data.weatherCurrent) {
-								data.weatherChange = `⚠️ ${translator.translate('Possible weather change at')} ${data.weatherChangeTime} : ➡️ ${translator.translate(this.GameData.utilData.weather[data.weatherNext].name)} ${translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherNext].emoji, platform))}`
-								data.weatherCurrentName = translator.translate('unknown')
-								data.weatherCurrentEmoji = '❓'
+					let pokemonShouldBeBoosted = false
+					let pokemonWillBeBoosted = false
+					const currentBoostedTypes = weatherForecast.current ? this.GameData.utilData.weatherTypeBoost[weatherForecast.current] : []
+					const forecastBoostedTypes = weatherForecast.next ? this.GameData.utilData.weatherTypeBoost[weatherForecast.next] : []
+					if (weatherForecast.current > 0 && currentBoostedTypes.filter((boostedType) => data.types.includes(boostedType)).length > 0) pokemonShouldBeBoosted = true
+					if (weatherForecast.next > 0 && ((data.weather > 0 && weatherForecast.next !== data.weather) || (weatherForecast.current > 0 && weatherForecast.next !== weatherForecast.current) || (pokemonShouldBeBoosted && data.weather === 0))) {
+						const weatherChangeTime = moment((data.disappear_time - (data.disappear_time % 3600)) * 1000)
+							.tz(geoTz.find(data.latitude, data.longitude)[0].toString())
+							.format(this.config.locale.time)
+							.slice(0, -3)
+						pokemonWillBeBoosted = forecastBoostedTypes.filter((boostedType) => data.types.includes(boostedType)).length > 0 ? 1 : 0
+						if (data.weather > 0 && !pokemonWillBeBoosted || data.weather === 0 && pokemonWillBeBoosted) {
+							weatherForecast.current = data.weather > 0 ? data.weather : weatherForecast.current
+							if (pokemonShouldBeBoosted && data.weather === 0) {
+								data.weatherCurrent = 0
 							} else {
-								data.weatherChange = `⚠️ ${translator.translate('Possible weather change at')} ${data.weatherChangeTime} : ${translator.translate(this.GameData.utilData.weather[data.weatherCurrent].name)} ${translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherCurrent].emoji, platform))} ➡️ ${translator.translate(this.GameData.utilData.weather[data.weatherNext].name)} ${translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherNext].emoji, platform))}`
-								data.weatherCurrentName = translator.translate(this.GameData.utilData.weather[data.weatherCurrent].name)
-								data.weatherCurrentEmoji = translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherCurrent].emoji, platform))
+								data.weatherCurrent = weatherForecast.current
 							}
-							data.weatherNextName = translator.translate(this.GameData.utilData.weather[data.weatherNext].name)
-							data.weatherNextEmoji = translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherNext].emoji, platform))
+							data.weatherChangeTime = weatherChangeTime
+							data.weatherNext = weatherForecast.next
 						}
+					}
+				}
 
-						const e = []
-						const n = []
-						monster.types.forEach((type) => {
-							e.push(translator.translate(this.emojiLookup.lookup(this.GameData.utilData.types[type.name].emoji, platform)))
-							n.push(type.name)
-						})
-						data.emoji = e
-						data.typeNameEng = n
-						data.typeName = data.typeNameEng.map((type) => translator.translate(type)).join(', ')
-						data.emojiString = e.join('')
-						data.typeEmoji = data.emojiString
+				// Future event processing
+				const event = this.eventParser.eventChangesSpawn(moment().unix(), data.disappear_time, data.latitude, data.longitude)
+				if (event) {
+					data.futureEvent = true
+					data.futureEventTime = event.time
+					data.futureEventName = event.name
+					data.futureEventTrigger = event.reason
+				}
 
-						require('./common/evolutionCalculator').setEvolutions(data, this.GameData, this.log, logReference, translator, this.emojiLookup, platform, monster)
+				// Lookup pokestop name if needed
+				if (this.config.general.populatePokestopName && !data.pokestopName && data.pokestop_id && this.scannerQuery) {
+					data.pokestopName = this.escapeJsonString(await this.scannerQuery.getPokestopName(data.pokestop_id))
+				}
 
-						const createPvpDisplay = (leagueCap, leagueData, maxRank, minCp) => {
-							const displayList = []
-							for (const rank of leagueData) {
-								if (rank.rank <= maxRank && rank.cp >= minCp) {
-									const displayRank = {}
-									displayRank.rank = +rank.rank
-									displayRank.formId = +rank.form || 0
-									displayRank.evolution = rank.evolution
-									displayRank.level = +rank.level
-									displayRank.cap = rank.cap
-									displayRank.capped = rank.capped
-									displayRank.levelWithCap = displayRank.cap && !displayRank.capped ? `${displayRank.level}/${displayRank.cap}` : displayRank.level
-									displayRank.cp = rank.cp
-									displayRank.pokemonId = +rank.pokemon
-									displayRank.percentage = rank.percentage <= 1 ? (rank.percentage * 100).toFixed(2) : rank.percentage.toFixed(2)
+				for (const cares of consolidatedAlerts) {
+					this.log.debug(`${logReference}: [matched] Creating monster alert for ${cares.id} ${cares.name} ${cares.type} ${cares.language} ${cares.template}`)
 
-									let monsterName
-									let formName
-									let stats
-									let mon = this.GameData.monsters[`${displayRank.pokemonId}_${displayRank.formId}`]
+					const rateLimitTtr = this.getRateLimitTimeToRelease(cares.id)
+					if (rateLimitTtr) {
+						this.log.verbose(`${logReference}: [matched] Not creating monster alert (Rate limit) for ${cares.type} ${cares.id} ${cares.name} Time to release: ${rateLimitTtr}`)
+						continue // eslint-disable-line no-continue
+					}
+
+					const language = cares.language || this.config.general.locale
+					const translator = this.translatorFactory.Translator(language)
+					let [platform] = cares.type.split(':')
+					if (platform === 'webhook') platform = 'discord'
+
+					data.name = translator.translate(monster.name)
+					data.formName = translator.translate(monster.form.name)
+					if (data.disguisePokemonNameEng) data.disguisePokemonName = translator.translate(data.disguisePokemonNameEng)
+					if (data.disguiseFormNameEng) data.disguiseFormName = translator.translate(data.disguiseFormNameEng)
+
+					data.formNormalisedEng = data.formNameEng === 'Normal' ? '' : data.formNameEng
+					data.formNormalised = translator.translate(data.formNormalisedEng)
+
+					data.fullNameEng = data.nameEng.concat(data.formNormalisedEng ? ' ' : '', data.formNormalisedEng)
+					data.fullName = data.name.concat(data.formNormalised ? ' ' : '', data.formNormalised)
+
+					data.genderData = {
+						name: translator.translate(data.genderDataEng.name),
+						emoji: translator.translate(this.emojiLookup.lookup(data.genderDataEng.emoji, platform)),
+					}
+					data.genderName = data.genderData.name
+					data.genderEmoji = data.genderData.emoji
+					data.shinyPossibleEmoji = data.shinyPossible ? translator.translate(this.emojiLookup.lookup('shiny', platform)) : ''
+					data.rarityName = translator.translate(data.rarityNameEng)
+					data.sizeName = translator.translate(data.sizeNameEng)
+					data.quickMoveName = encountered && this.GameData.moves[data.quickMoveId] ? translator.translate(this.GameData.moves[data.quickMoveId].name) : ''
+					data.quickMoveEmoji = this.GameData.moves[data.quickMoveId] && this.GameData.utilData.types[this.GameData.moves[data.quickMoveId].type] ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.types[this.GameData.moves[data.quickMoveId].type].emoji, platform)) : ''
+					data.chargeMoveName = encountered && this.GameData.moves[data.chargeMoveId] ? translator.translate(this.GameData.moves[data.chargeMoveId].name) : ''
+					data.chargeMoveEmoji = this.GameData.moves[data.chargeMoveId] && this.GameData.utilData.types[this.GameData.moves[data.chargeMoveId].type] ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.types[this.GameData.moves[data.chargeMoveId].type].emoji, platform)) : ''
+					data.boosted = !!data.weather
+					data.generationName = translator.translate(data.generationNameEng)
+					data.boostWeatherId = data.weather ? data.weather : ''
+					data.boostWeatherName = data.weather ? translator.translate(this.GameData.utilData.weather[data.weather].name) : ''
+					data.boostWeatherEmoji = data.weather ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weather].emoji, platform)) : ''
+					data.gameWeatherId = this.GameData.utilData.weather[currentCellWeather] ? currentCellWeather : ''
+					data.gameWeatherName = this.GameData.utilData.weather[currentCellWeather] ? translator.translate(this.GameData.utilData.weather[currentCellWeather].name) : ''
+					data.gameWeatherEmoji = this.GameData.utilData.weather[currentCellWeather] ? translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[currentCellWeather].emoji, platform)) : ''
+					data.formname = data.formName // deprecated
+					data.quickMove = data.quickMoveName // deprecated
+					data.chargeMove = data.chargeMoveName // deprecated
+					data.move1emoji = data.quickMoveEmoji // deprecated
+					data.move2emoji = data.chargeMoveEmoji // deprecated
+					data.boost = data.boostWeatherName // deprecated
+					data.boostemoji = data.boostWeatherEmoji // deprecated
+					data.gameweather = data.gameWeatherName // deprecated
+					data.gameweatheremoji = data.gameWeatherEmoji // deprecated
+
+					if (data.weatherNext) {
+						if (!data.weatherCurrent) {
+							data.weatherChange = `⚠️ ${translator.translate('Possible weather change at')} ${data.weatherChangeTime} : ➡️ ${translator.translate(this.GameData.utilData.weather[data.weatherNext].name)} ${translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherNext].emoji, platform))}`
+							data.weatherCurrentName = translator.translate('unknown')
+							data.weatherCurrentEmoji = '❓'
+						} else {
+							data.weatherChange = `⚠️ ${translator.translate('Possible weather change at')} ${data.weatherChangeTime} : ${translator.translate(this.GameData.utilData.weather[data.weatherCurrent].name)} ${translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherCurrent].emoji, platform))} ➡️ ${translator.translate(this.GameData.utilData.weather[data.weatherNext].name)} ${translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherNext].emoji, platform))}`
+							data.weatherCurrentName = translator.translate(this.GameData.utilData.weather[data.weatherCurrent].name)
+							data.weatherCurrentEmoji = translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherCurrent].emoji, platform))
+						}
+						data.weatherNextName = translator.translate(this.GameData.utilData.weather[data.weatherNext].name)
+						data.weatherNextEmoji = translator.translate(this.emojiLookup.lookup(this.GameData.utilData.weather[data.weatherNext].emoji, platform))
+					}
+
+					const e = []
+					const n = []
+					monster.types.forEach((type) => {
+						e.push(translator.translate(this.emojiLookup.lookup(this.GameData.utilData.types[type.name].emoji, platform)))
+						n.push(type.name)
+					})
+					data.emoji = e
+					data.typeNameEng = n
+					data.typeName = data.typeNameEng.map((type) => translator.translate(type)).join(', ')
+					data.emojiString = e.join('')
+					data.typeEmoji = data.emojiString
+
+					require('./common/evolutionCalculator').setEvolutions(data, this.GameData, this.log, logReference, translator, this.emojiLookup, platform, monster)
+
+					const createPvpDisplay = (leagueCap, leagueData, maxRank, minCp) => {
+						const displayList = []
+						for (const rank of leagueData) {
+							if (rank.rank <= maxRank && rank.cp >= minCp) {
+								const displayRank = {}
+								displayRank.rank = +rank.rank
+								displayRank.formId = +rank.form || 0
+								displayRank.evolution = rank.evolution
+								displayRank.level = +rank.level
+								displayRank.cap = rank.cap
+								displayRank.capped = rank.capped
+								displayRank.levelWithCap = displayRank.cap && !displayRank.capped ? `${displayRank.level}/${displayRank.cap}` : displayRank.level
+								displayRank.cp = rank.cp
+								displayRank.pokemonId = +rank.pokemon
+								displayRank.percentage = rank.percentage <= 1 ? (rank.percentage * 100).toFixed(2) : rank.percentage.toFixed(2)
+
+								let monsterName
+								let formName
+								let stats
+								let mon = this.GameData.monsters[`${displayRank.pokemonId}_${displayRank.formId}`]
+								if (!mon) {
+									mon = this.GameData.monsters[`${displayRank.pokemonId}_0`]
 									if (!mon) {
-										mon = this.GameData.monsters[`${displayRank.pokemonId}_0`]
-										if (!mon) {
-											monsterName = `${translator.translate('Unknown monster')} ${displayRank.pokemonId}`
-											stats = { baseAttack: 0, baseDefense: 0, baseStamina: 0 }
-										} else {
-											monsterName = mon.name
-											stats = monster.stats
-										}
-										formName = `${displayRank.formId}`
+										monsterName = `${translator.translate('Unknown monster')} ${displayRank.pokemonId}`
+										stats = { baseAttack: 0, baseDefense: 0, baseStamina: 0 }
 									} else {
 										monsterName = mon.name
-										formName = mon.form.name
-										if (formName === undefined || formName === 'Normal') formName = ''
 										stats = monster.stats
 									}
-									displayRank.baseStats = stats
-									displayRank.nameEng = monsterName
-									displayRank.formEng = formName
-									displayRank.name = translator.translate(monsterName)
-									displayRank.form = translator.translate(formName)
-									if (displayRank.evolution) {
-										displayRank.fullNameEng = translator.format(this.GameData.utilData.megaName[displayRank.evolution], displayRank.nameEng.concat(displayRank.formEng ? ' ' : '', displayRank.formEng))
-										displayRank.fullName = translator.translateFormat(this.GameData.utilData.megaName[displayRank.evolution], displayRank.name.concat(displayRank.form ? ' ' : '', displayRank.form))
-									} else {
-										displayRank.fullNameEng = displayRank.nameEng.concat(displayRank.formEng ? ' ' : '', displayRank.formEng)
-										displayRank.fullName = displayRank.name.concat(displayRank.form ? ' ' : '', displayRank.form)
-									}
+									formName = `${displayRank.formId}`
+								} else {
+									monsterName = mon.name
+									formName = mon.form.name
+									if (formName === undefined || formName === 'Normal') formName = ''
+									stats = monster.stats
+								}
+								displayRank.baseStats = stats
+								displayRank.nameEng = monsterName
+								displayRank.formEng = formName
+								displayRank.name = translator.translate(monsterName)
+								displayRank.form = translator.translate(formName)
+								if (displayRank.evolution) {
+									displayRank.fullNameEng = translator.format(this.GameData.utilData.megaName[displayRank.evolution], displayRank.nameEng.concat(displayRank.formEng ? ' ' : '', displayRank.formEng))
+									displayRank.fullName = translator.translateFormat(this.GameData.utilData.megaName[displayRank.evolution], displayRank.name.concat(displayRank.form ? ' ' : '', displayRank.form))
+								} else {
+									displayRank.fullNameEng = displayRank.nameEng.concat(displayRank.formEng ? ' ' : '', displayRank.formEng)
+									displayRank.fullName = displayRank.name.concat(displayRank.form ? ' ' : '', displayRank.form)
+								}
 
-									displayRank.matchesUserTrack = false
-									if (cares.filters.length) {
-										displayRank.passesFilter = false
-										for (const filter of cares.filters) {
-											if ((filter.pvp_ranking_league === leagueCap || filter.pvp_ranking_league === 0)
+								displayRank.matchesUserTrack = false
+								if (cares.filters.length) {
+									displayRank.passesFilter = false
+									for (const filter of cares.filters) {
+										if ((filter.pvp_ranking_league === leagueCap || filter.pvp_ranking_league === 0)
 											&& (filter.pvp_ranking_cap === 0 || filter.pvp_ranking_cap === displayRank.cap || displayRank.capped)
 											&& (filter.pvp_ranking_worst >= displayRank.rank)) {
-												displayRank.passesFilter = true
-												displayRank.matchesUserTrack = true
-											}
+											displayRank.passesFilter = true
+											displayRank.matchesUserTrack = true
 										}
-									} else {
-										displayRank.passesFilter = true
 									}
-									if (!this.config.pvp.filterByTrack || displayRank.passesFilter) {
-										displayList.push(displayRank)
-									}
+								} else {
+									displayRank.passesFilter = true
+								}
+								if (!this.config.pvp.filterByTrack || displayRank.passesFilter) {
+									displayList.push(displayRank)
 								}
 							}
-							return displayList.length ? displayList : null
 						}
-
-						const calculateBestInfo = (ranks) => {
-							if (!ranks) return null
-							const best = { rank: 4096, list: [] }
-							for (const result of ranks) {
-								if (result.rank === best.rank) {
-									best.list.push(result)
-								} else if (result.rank < best.rank) {
-									best.rank = result.rank
-									best.list = [result]
-								}
-							}
-							best.name = Array.from(new Set(best.list.map((x) => x.fullName))).join(', ')
-							best.nameEng = Array.from(new Set(best.list.map((x) => x.fullNameEng))).join(', ')
-							return best
-						}
-
-						data.pvpGreat = data.pvp_rankings_great_league ? createPvpDisplay(1500, data.pvp_rankings_great_league, this.config.pvp.pvpDisplayMaxRank, this.config.pvp.pvpDisplayGreatMinCP) : null
-						data.pvpGreatBest = calculateBestInfo(data.pvpGreat)
-						data.pvpUltra = data.pvp_rankings_ultra_league ? createPvpDisplay(2500, data.pvp_rankings_ultra_league, this.config.pvp.pvpDisplayMaxRank, this.config.pvp.pvpDisplayUltraMinCP) : null
-						data.pvpUltraBest = calculateBestInfo(data.pvpUltra)
-						data.pvpLittle = data.pvp_rankings_little_league ? createPvpDisplay(500, data.pvp_rankings_little_league, this.config.pvp.pvpDisplayMaxRank, this.config.pvp.pvpDisplayLittleMinCP) : null
-						data.pvpLittleBest = calculateBestInfo(data.pvpLittle)
-						data.pvpAvailable = data.pvpGreat !== null || data.pvpUltra !== null || data.pvpLittle !== null
-						data.userHasPvpTracks = !!cares.filters.length
-
-						data.distance = cares.longitude ? this.getDistance({ lat: cares.latitude, lon: cares.longitude }, { lat: data.latitude, lon: data.longitude }) : ''
-
-						const bearing = cares.longitude ? this.getBearing({ lat: cares.latitude, lon: cares.longitude }, { lat: data.latitude, lon: data.longitude }) : null
-						data.bearing = bearing?.toFixed(0) ?? ''
-						data.bearingEmoji = bearing ? this.emojiLookup.lookup(this.getBearingEmoji(data.bearing), platform) : ''
-
-						const view = {
-							...geoResult,
-							...data,
-							id: data.pokemon_id,
-							baseStats: monster.stats,
-							time: data.disappearTime,
-							tthh: data.tth.hours,
-							tthm: data.tth.minutes,
-							tths: data.tth.seconds,
-							now: new Date(),
-							nowISO: new Date().toISOString(),
-							pvpUserRanking: cares.pvp_ranking_worst === 4096 ? 0 : cares.pvp_ranking_worst,
-							areas: data.matchedAreas.filter((area) => area.displayInMatches).map((area) => area.name).join(', '),
-							pvpDisplayMaxRank: this.config.pvp.pvpDisplayMaxRank,
-							pvpDisplayGreatMinCP: this.config.pvp.pvpDisplayGreatMinCP,
-							pvpDisplayUltraMinCP: this.config.pvp.pvpDisplayUltraMinCP,
-							pvpDisplayLittleMinCP: this.config.pvp.pvpDisplayLittleMinCP,
-						}
-
-						const templateType = (data.iv === -1) ? 'monsterNoIv' : 'monster'
-						const message = await this.createMessage(logReference, templateType, platform, cares.template, language, cares.ping, view)
-
-						const work = {
-							lat: data.latitude.toString().substring(0, 8),
-							lon: data.longitude.toString().substring(0, 8),
-							message,
-							target: cares.id,
-							type: cares.type,
-							name: cares.name,
-							tth: data.tth,
-							clean: cares.clean,
-							emoji: data.emoji,
-							logReference,
-							language,
-						}
-						jobs.push(work)
+						return displayList.length ? displayList : null
 					}
 
-					this.emit('postMessage', jobs)
-				} catch (e) {
-					this.log.error(`${data.encounter_id}: [matched] Can't seem to handle monster (user cared): `, e, data)
+					const calculateBestInfo = (ranks) => {
+						if (!ranks) return null
+						const best = { rank: 4096, list: [] }
+						for (const result of ranks) {
+							if (result.rank === best.rank) {
+								best.list.push(result)
+							} else if (result.rank < best.rank) {
+								best.rank = result.rank
+								best.list = [result]
+							}
+						}
+						best.name = Array.from(new Set(best.list.map((x) => x.fullName))).join(', ')
+						best.nameEng = Array.from(new Set(best.list.map((x) => x.fullNameEng))).join(', ')
+						return best
+					}
+
+					data.pvpGreat = data.pvp_rankings_great_league ? createPvpDisplay(1500, data.pvp_rankings_great_league, this.config.pvp.pvpDisplayMaxRank, this.config.pvp.pvpDisplayGreatMinCP) : null
+					data.pvpGreatBest = calculateBestInfo(data.pvpGreat)
+					data.pvpUltra = data.pvp_rankings_ultra_league ? createPvpDisplay(2500, data.pvp_rankings_ultra_league, this.config.pvp.pvpDisplayMaxRank, this.config.pvp.pvpDisplayUltraMinCP) : null
+					data.pvpUltraBest = calculateBestInfo(data.pvpUltra)
+					data.pvpLittle = data.pvp_rankings_little_league ? createPvpDisplay(500, data.pvp_rankings_little_league, this.config.pvp.pvpDisplayMaxRank, this.config.pvp.pvpDisplayLittleMinCP) : null
+					data.pvpLittleBest = calculateBestInfo(data.pvpLittle)
+					data.pvpAvailable = data.pvpGreat !== null || data.pvpUltra !== null || data.pvpLittle !== null
+					data.userHasPvpTracks = !!cares.filters.length
+
+					data.distance = cares.longitude ? this.getDistance({ lat: cares.latitude, lon: cares.longitude }, { lat: data.latitude, lon: data.longitude }) : ''
+
+					const bearing = cares.longitude ? this.getBearing({ lat: cares.latitude, lon: cares.longitude }, { lat: data.latitude, lon: data.longitude }) : null
+					data.bearing = bearing?.toFixed(0) ?? ''
+					data.bearingEmoji = bearing ? this.emojiLookup.lookup(this.getBearingEmoji(data.bearing), platform) : ''
+
+					const view = {
+						...geoResult,
+						...data,
+						id: data.pokemon_id,
+						baseStats: monster.stats,
+						time: data.disappearTime,
+						tthh: data.tth.hours,
+						tthm: data.tth.minutes,
+						tths: data.tth.seconds,
+						now: new Date(),
+						nowISO: new Date().toISOString(),
+						pvpUserRanking: cares.pvp_ranking_worst === 4096 ? 0 : cares.pvp_ranking_worst,
+						areas: data.matchedAreas.filter((area) => area.displayInMatches).map((area) => area.name).join(', '),
+						pvpDisplayMaxRank: this.config.pvp.pvpDisplayMaxRank,
+						pvpDisplayGreatMinCP: this.config.pvp.pvpDisplayGreatMinCP,
+						pvpDisplayUltraMinCP: this.config.pvp.pvpDisplayUltraMinCP,
+						pvpDisplayLittleMinCP: this.config.pvp.pvpDisplayLittleMinCP,
+					}
+
+					const templateType = (data.iv === -1) ? 'monsterNoIv' : 'monster'
+					const message = await this.createMessage(logReference, templateType, platform, cares.template, language, cares.ping, view)
+
+					const work = {
+						lat: data.latitude.toString().substring(0, 8),
+						lon: data.longitude.toString().substring(0, 8),
+						message,
+						target: cares.id,
+						type: cares.type,
+						name: cares.name,
+						tth: data.tth,
+						clean: cares.clean,
+						emoji: data.emoji,
+						logReference,
+						language,
+					}
+					jobs.push(work)
 				}
-			})
-			return []
+
+				return jobs
+			} catch (e) {
+				this.log.error(`${data.encounter_id}: [matched] Can't seem to handle monster (user cared): `, e, data)
+				return []
+			}
 		} catch (e) {
 			this.log.error(`${data.encounter_id}: [matched] Can't seem to handle monster: `, e, data)
 		}
