+++ alerter/src/lib/poracleMessage/commands/poracle-test.js
@@ -151,10 +151,10 @@ exports.run = async (client, msg, args, options) => {
 				id: target.id,
 				name: target.name,
 				type: target.type,
-				language: language,
+				language,
 				latitude: human.latitude,
 				longitude: human.longitude,
-				template: template,
+				template,
 				clean: false,
 				ping: '',
 				distance: 0,
